+++ Stdfun.h
@@ -6,4 +6,4 @@
 
 void resetEncoders(Encoder *eREncoder, Encoder *eLEncoder);
 //void addToGlobalXPos(Encoder eREncoder, Encoder eREncoder, float ffXPosition);
-float returnDistance(Encoder *eEncoder);
+double returnDistance(Encoder *eEncoder);
